Completely new implementation for enemy spawning.
Every wave is randomly decided based on some inherent chance within an Enemy Wave Config. If and enemy or path is chosen that has been present in the last three waves, then it is run again. Once there is no conflicts, the wave is spawned. Weaker enemies have higher chance to spawn early on, and vice versa for stronger enemies. These chances change as the game moves along.